requirements.txt
numpy==1.7.1
#pyopencv>=2.0
pillow==2.2.1
pyttsx==1.1
pyttk>=0.3.2
pyaudio>=0.2.7
